Refactor schema operations
